3.2-crud/stocks_products/requirements.txt
django
djangorestframework
django_filters
psycopg2-binary
